requirements.txt
langchain
langchain-community
langchain-core
langchain-text-splitters
faiss-cpu
sentence-transformers
pypdf
python-docx
streamlit
groq
docx2txt
python-dotenv
flask_cors